+++ pokedex/src/App.js
@@ -1,5 +1,5 @@
+import { Router } from './routes/Router';
 import './App.css';
-import { Router } from './components/routes/Router';
 
 function App() {
   return (
