+++ pokedex/src/assets/img/pokelogo.png
