--- pokedex/src/components/Pokedex/styled.js
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const Header = styled.header`
-display: flex;
-justify-content: space-around;
-align-items: center;
-`
