+++ pokedex/src/constants/urls.js
@@ -1,0 +1,1 @@
+export const BASE_URL = "https://pokeapi.co/api/v2/"
