+++ pokedex/src/pages/DetailsPage/styled.js
