+++ pokedex/src/pages/Pokedex/styled.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const Header = styled.header`
+    display: flex;
+    height: 160px;
+    border: 1px solid black;
+    width: 100%;
+    justify-content: space-around;
+    align-items: center;
+`
