+++ pokedex/src/routes/Router.js
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { Home } from "../Home/Home";
-import { Pokedex } from "../Pokedex/Pokedex";
-import { DetailsPage } from "../DetailsPage/DetailsPage";
+import { Home } from "../pages/Home/Home";
+import { Pokedex } from "../pages/Pokedex/Pokedex";
+import { DetailsPage } from "../pages/DetailsPage/DetailsPage";
 
 export const Router = () => {
     return (
